+++ include/ignition/common/URI.hh
@@ -18,6 +18,7 @@
 #define IGNITION_COMMON_URI_HH_
 
 #include <memory>
+#include <optional>
 #include <string>
 
 #include <ignition/common/Export.hh>
@@ -28,10 +29,124 @@ namespace ignition
   namespace common
   {
     // Forward declare private data classes.
+    class URIAuthorityPrivate;
     class URIPathPrivate;
     class URIQueryPrivate;
     class URIFragmentPrivate;
     class URIPrivate;
+
+    /// A URI authority contains userinfo, host, and port data. The format
+    /// of a URI authority is `//userinfo@host:port`. The `userinfo` and
+    /// `port` components are optional.
+    ///
+    /// A URI Authority requires the existence of host information,
+    /// except when a Scheme is `file`.  When a scheme is `file`, then the
+    /// following are considered valid URIs and equivalent:
+    ///
+    ///   * file:/abs/path
+    ///   * file:///abs/path
+    ///
+    /// Keep in mind that a URI path must start with a forward slash when an
+    /// authority, as indicated by two forward slashes, is present. This
+    /// means relative file paths cannot be specified with an empty
+    /// authority. For example, `file://abs/path` will result in a host
+    /// value of `abs` and the URI path will be `/path`. You can specify
+    /// a relative path using `file:abs/path`.
+    class IGNITION_COMMON_VISIBLE URIAuthority
+    {
+      /// \brief Constructor
+      public: URIAuthority();
+
+      /// \brief Copy constructor.
+      /// \param[in] _path Another URIAuthority.
+      public: URIAuthority(const URIAuthority &_path);
+
+      /// \brief Construct a URIAuthority object from a string.
+      /// \param[in] _str A string.
+      public: explicit URIAuthority(const std::string &_str);
+
+      /// \brief Destructor
+      public: virtual ~URIAuthority();
+
+      /// \brief Remove all parts of the authority
+      public: void Clear();
+
+      /// \brief Get the user information.
+      /// \return User information, or an empty string if the user
+      /// information is not set.
+      public: std::string UserInfo() const;
+
+      /// \brief Set the user information.
+      /// \param[in] _userInfo The user information string.
+      public: void SetUserInfo(const std::string &_userInfo) const;
+
+      /// \brief Get the host.
+      /// \return The host.
+      public: std::string Host() const;
+
+      /// \brief Set the host.
+      /// \param[in] _host The host.
+      public: void SetHost(const std::string &_host) const;
+
+      /// \brief True if an empty host is considered valid.
+      /// \return True if an empty host is valid.
+      public: bool EmptyHostValid() const;
+
+      /// \brief Set whether an empty host is considered valid.
+      /// This should only be set to true if the corresponding URIScheme
+      /// is "file".
+      /// \param[in] _valid True if an empty host is valid.
+      public: void SetEmptyHostValid(bool _valid) const;
+
+      /// \brief Get the port.
+      /// \return The port number, which is optional.
+      public: std::optional<int> Port() const;
+
+      /// \brief Set the port number.
+      public: void SetPort(int _port) const;
+
+      /// \brief Return true if the two authorities match.
+      /// \param[in] _auth Authority.
+      /// return True of the authorities match.
+      public: bool operator==(const URIAuthority &_auth) const;
+
+      /// \brief Get the complete authoriy as a string.
+      /// \return The authority as a string.
+      public: std::string Str() const;
+
+      /// \brief Assignment operator.
+      /// \param[in] _auth Another URIAuthority.
+      /// \return Itself.
+      public: URIAuthority &operator=(const URIAuthority &_auth);
+
+      /// \brief Return true if the string is a valid path.
+      /// \param[in] _str String to check.
+      /// \param[in] _emptyHostValid Set this to true if an empty host is
+      /// valid. This should only be set to true if the corresponding URIScheme
+      /// is "file".
+      /// \return True if _str is a valid URI path.
+      public: static bool Valid(const std::string &_str,
+                  bool _emptyHostValid = false);
+
+      /// \brief Return true if this is a valid authority.
+      /// \return True if this is a valid URI authority.
+      public: bool Valid() const;
+
+      /// \brief Parse a string as URIAuthority.
+      /// \param[in] _str A string.
+      /// \param[in] _emptyHostValid Set this to true if an empty host is
+      /// valid. This should only be set to true if the corresponding URIScheme
+      /// is "file".
+      /// \return True if the string could be parsed as a URIAuthority.
+      public: bool Parse(const std::string &_str,
+                  bool _emptyHostValid = false);
+
+      IGN_COMMON_WARN_IGNORE__DLL_INTERFACE_MISSING
+      /// \internal
+      /// \brief Pointer to private data.
+      private: std::unique_ptr<URIAuthorityPrivate> dataPtr;
+      IGN_COMMON_WARN_RESUME__DLL_INTERFACE_MISSING
+    };
 
     /// \brief The path component of a URI
     class IGNITION_COMMON_VISIBLE URIPath
@@ -258,7 +373,10 @@ namespace ignition
       IGN_COMMON_WARN_RESUME__DLL_INTERFACE_MISSING
     };
 
-    /// \brief A complete URI
+    /// \brief A complete URI which has the following components:
+    ///
+    /// scheme:[//authority]path[?query][#fragment]
+    ///
     // cppcheck-suppress class_X_Y
     class IGNITION_COMMON_VISIBLE URI
     {
@@ -278,7 +396,7 @@ namespace ignition
 
       /// \brief Get the URI as a string, which has the form:
       ///
-      /// scheme://path?query
+      /// scheme:[//authority]path[?query][#fragment]
       ///
       /// \return The full URI as a string
       public: std::string Str() const;
@@ -293,6 +411,14 @@ namespace ignition
       /// \brief Set the URI's scheme
       /// \param[in] _scheme New scheme.
       public: void SetScheme(const std::string &_scheme);
+
+      /// \brief Get a mutable version of the URI's authority.
+      /// \return The authority
+      public: URIAuthority &Authority();
+
+      /// \brief Get a const reference of the URI's authority.
+      /// \return The authority
+      public: const URIAuthority &Authority() const;
 
       /// \brief Get a mutable version of the path component
       /// \return A reference to the path